+++ lib/ax100_decode_impl.h
@@ -1,21 +1,11 @@
 /* -*- c++ -*- */
 /* 
- * Copyright 2016 Daniel Estevez <daniel@destevez.net>.
- * 
- * This is free software; you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation; either version 3, or (at your option)
- * any later version.
- * 
- * This software is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- * 
- * You should have received a copy of the GNU General Public License
- * along with this software; see the file COPYING.  If not, write to
- * the Free Software Foundation, Inc., 51 Franklin Street,
- * Boston, MA 02110-1301, USA.
+ * Copyright 2016,2020 Daniel Estevez <daniel@destevez.net>.
+ *
+ * This file is part of gr-satellites
+ *
+ * SPDX-License-Identifier: GPL-3.0-or-later
+ *
  */
 
 #ifndef INCLUDED_SATELLITES_AX100_DECODE_IMPL_H
@@ -23,13 +13,16 @@
 
 #include <satellites/ax100_decode.h>
 
+#include <array>
+
 namespace gr {
   namespace satellites {
 
     class ax100_decode_impl : public ax100_decode
     {
      private:
-      bool d_verbose;
+      std::array<uint8_t, 256> d_data;
+      const bool d_verbose;
 
      public:
       ax100_decode_impl(bool verbose);