+++ python/telemetry/amicalsat.py
@@ -1,0 +1,19 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+
+# Copyright 2020 Daniel Estevez <daniel@destevez.net>
+#
+# This file is part of gr-satellites
+#
+# SPDX-License-Identifier: GPL-3.0-or-later
+# 
+
+from construct import *
+
+from .ax25 import Header
+
+amicalsat = Struct(
+        'header' / Header,
+        'info' / GreedyString('ascii')
+        )
+